Move base for plugin support into codyze-core
