+++ codyze-core/src/main/kotlin/de/fraunhofer/aisec/codyze/core/plugin/Plugin.kt
@@ -13,13 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.fraunhofer.aisec.codyze.plugins
+package de.fraunhofer.aisec.codyze.core.plugin
 
 import com.github.ajalt.clikt.core.NoOpCliktCommand
 import com.github.ajalt.clikt.parameters.groups.provideDelegate
 import de.fraunhofer.aisec.codyze.core.output.aggregator.Aggregate
 import de.fraunhofer.aisec.codyze.core.output.aggregator.extractLastRun
 import io.github.oshai.kotlinlogging.KotlinLogging
+import org.koin.core.module.Module
 import java.io.File
 import java.nio.file.Path
 
@@ -32,7 +33,8 @@ val logger = KotlinLogging.logger { }
  * otherwise it will not be selectable in the configuration.
  * Also, remember to add a page to docs/plugins.
  */
-abstract class Plugin(private val cliName: String) : NoOpCliktCommand(hidden = true, name = cliName.lowercase()) {
+abstract class Plugin(private val cliName: String) :
+    NoOpCliktCommand(hidden = true, name = cliName.lowercase()) {
     private val options by PluginOptionGroup(cliName)
 
     /**
@@ -42,6 +44,8 @@ abstract class Plugin(private val cliName: String) : NoOpCliktCommand(hidden = t
      * @param output The location of the results
      */
     abstract fun execute(target: List<Path>, context: List<Path>, output: File)
+
+    abstract fun module(): Module
 
     /**
      * Define two plugins as equal if they are of the same type and therefore have the same CLI name.
