+++ codyze-core/src/main/kotlin/de/fraunhofer/aisec/codyze/core/plugin/PluginOptionGroup.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.fraunhofer.aisec.codyze.plugins
+package de.fraunhofer.aisec.codyze.core.plugin
 
 import com.github.ajalt.clikt.parameters.groups.OptionGroup
 import com.github.ajalt.clikt.parameters.options.default
